structure components on folders
