+++ src/__tests__/components/Register.test.js
@@ -10,7 +10,7 @@ import userEvent from "@testing-library/user-event";
 import { rest } from "msw";
 import { server } from "../../mocks/server.js";
 
-import Register from "../../pages/Register";
+import Register from "../../components/Register";
 
 import { act } from "react-dom/test-utils";
 import { API_URL } from "../../services/api.js";
