+++ src/components/Sidebar/useFetchTags.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
-import { api } from "../services/apiClient";
-import { formatedErrorsArray } from "../utils/request";
+import { api } from "../../services/apiClient";
+import { formatedErrorsArray } from "../../utils/request";
 
 export const useFetchTags = () => {
   const [tags, setTags] = useState([]);
